first updates in general project
