--- SeleniumTraining_DemoWebshop_Sampleapp/src/main/resources/step_definition/demo_webshop/login/homeStepDefinitions.java
@@ -1,8 +1,0 @@
-package step_definition.demo_webshop.login;
-
-public class homeStepDefinitions {
-    public homeStepDefinitions() {
-        And("^I click on computers on the top menu$", () -> {
-        });
-    }
-}
