+++ SeleniumTraining_DemoWebshop_Sampleapp/src/main/resources/step_definition/login/homeStepDefinitions.java
@@ -1,0 +1,8 @@
+package step_definition.login;
+
+public class homeStepDefinitions {
+    public homeStepDefinitions() {
+        And("^I click on computers on the top menu$", () -> {
+        });
+    }
+}
