+++ SeleniumTraining_DemoWebshop_Sampleapp/src/test/java/actions/cart_action/CartAction.java
@@ -1,0 +1,11 @@
+package actions.cart_action;
+
+import pages.cart_page.CartPage;
+
+public class CartAction extends CartPage {
+
+    public void clickOnDesktopOption(){
+        genericMethods.validateVisibilityOfElement(desktopOption);
+        genericMethods.click(desktopOption);
+    }
+}
