--- SeleniumTraining_DemoWebshop_Sampleapp/src/test/java/actions/demo_webshop/cart_action/CartAction.java
@@ -1,11 +1,0 @@
-package actions.demo_webshop.cart_action;
-
-import pages.demo_webshop.cart_page.CartPage;
-
-public class CartAction extends CartPage {
-
-    public void clickOnDesktopOption(){
-        genericMethods.validateVisibilityOfElement(desktopOption);
-        genericMethods.click(desktopOption);
-    }
-}
