+++ SeleniumTraining_DemoWebshop_Sampleapp/src/test/java/actions/home_action/HomeAction.java
@@ -1,0 +1,27 @@
+package actions.home_action;
+
+import generic_methods.GenericMethods;
+import pages.home_page.HomePage;
+
+public class HomeAction extends HomePage {
+
+    public GenericMethods genericMethods = new GenericMethods();
+
+    public void clickOnLoginLink() {
+        // Click on login link
+        genericMethods.click(loginLink);
+
+        // Validate Login Screen Visibility
+        genericMethods.validateVisibilityOfElement(validateLoginScreen);
+
+    }
+
+    public void sucessLoginValidation() {
+        genericMethods.validateVisibilityOfElement(logoutButton);
+    }
+
+    public void selectComputerOptionOnTopMenu(){
+        genericMethods.validateVisibilityOfElement(topMenuComputers);
+        genericMethods.click(topMenuComputers);
+    }
+}
