+++ SeleniumTraining_DemoWebshop_Sampleapp/src/test/java/pages/cart_page/CartPage.java
@@ -1,0 +1,11 @@
+package pages.cart_page;
+
+import base.BasePage;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+
+public class CartPage extends BasePage {
+
+    @FindBy(xpath = "//div[@class ='sub-category-grid']//a[contains(@title, 'Desktops')]/parent::div[@class = 'picture']")
+    protected WebElement desktopOption;
+}
